Fixed several fuck-ups of mine Last time i accidentally modified the lenovo scraper when i meant to change the newegg one. Moved these changes to the right scraper.
The categories block in pipelines.py doesn't work when it's a new
category.  Added handler for all conditions (0, 1, and many found).

Updated lenovo scraper.

+++ apps/scrapy/pipelines.py
@@ -194,20 +194,23 @@ class CategoryPipeline(object):
         # Should dissociate all the products from all the different categories
         # and associate them with the "good" one (whichever has lowest id),
         # then delete the "bad" category.
-        try:
-            category, _ = Category.objects.get_or_create(**kwargs)
-        except MultipleObjectsReturned:
+
+        categories = Category.objects.filter(**kwargs)
+        if len(categories) > 1:
             print "Killing categories"
-            categories = Category.objects.filter(**kwargs).order_by('id')
             category = categories[0]
-            category.name = category.name.lower()
             print "keeping category", category.name
             for duplicate_category in categories[1:]:
                 print "\tkilling:", duplicate_category.name
                 category.products.add(*duplicate_category.products.all())
                 print "\ttransferred", duplicate_category.products.all(), "products"
                 duplicate_category.delete()
+        elif len(categories) == 1:
+            category = categories[0]
+        else:
+            category = Category.objects.create(store=kwargs['store'], name=name.lower())
 
+        category.name = category.name.lower()
         category.save()
 
         try:
